Here I have worked to cut down on the number of warning messages produced using g++ 3.1 with the options -ansi -pedantic -Wall.  Most of the warnings were minor.  Most of the warnings (just 20 in the build I did) relate to unused variables and comparing signed and unsigned intergers.  The former is a minor minor performance and maintainability issue while the latter can cause real problems when things go negative.  I have build a debug (-g -O0 -D_DEBUG) and release (-O2) versions and ran the checkin test suite on both.  The tests for TSFCore, Teuchos, Epetra, AztecOO and Ifpack all passed but the tests for NOX and EpetraExt failed.  I do not know how I could have broken NOX or EpetraExt, perhaps the tests for these packages are just broken.

+++ packages/TSFCore/src/adapters/aztecoo/Epetra_ProductOperator.hpp
@@ -403,7 +403,7 @@ inline
 void ProductOperator::validateIndex(int k) const
 {
 	TEST_FOR_EXCEPTION(
-		k < 0 || Op_.size()-1 < k, std::logic_error
+		k < 0 || static_cast<int>(Op_.size())-1 < k, std::logic_error
 		,"Epetra::ProductOperator: Error, k = "<<k<< " is not in the range [0,"<<Op_.size()-1<<"]!"
 		);
 }
