+++ packages/TSFCore/src/adapters/mpi-base/TSFCoreMPIVectorBaseDecl.hpp
@@ -157,9 +157,9 @@ public:
 	 */
 	void applyOp(
 		const RTOpPack::RTOpT<Scalar>   &op
-		,const size_t                   num_vecs
+		,const int                   num_vecs
 		,const Vector<Scalar>*          vecs[]
-		,const size_t                   num_targ_vecs
+		,const int                   num_targ_vecs
 		,Vector<Scalar>*                targ_vecs[]
 		,RTOpPack::ReductTarget         *reduct_obj
 		,const Index                    first_ele
