+++ packages/TSFCore/src/interfaces/Core/TSFCoreMultiVectorDecl.hpp
@@ -298,9 +298,9 @@ public:
 	 */
 	virtual void applyOp(
 		const RTOpPack::RTOpT<Scalar>   &primary_op
-		,const size_t                   num_multi_vecs
+		,const int                   num_multi_vecs
 		,const MultiVector<Scalar>*     multi_vecs[]
-		,const size_t                   num_targ_multi_vecs
+		,const int                   num_targ_multi_vecs
 		,MultiVector<Scalar>*           targ_multi_vecs[]
 		,RTOpPack::ReductTarget*        reduct_objs[]
 		,const Index                    primary_first_ele
@@ -334,9 +334,9 @@ public:
 	virtual void applyOp(
 		const RTOpPack::RTOpT<Scalar>   &primary_op
 		,const RTOpPack::RTOpT<Scalar>  &secondary_op
-		,const size_t                   num_multi_vecs
+		,const int                   num_multi_vecs
 		,const MultiVector<Scalar>*     multi_vecs[]
-		,const size_t                   num_targ_multi_vecs
+		,const int                   num_targ_multi_vecs
 		,MultiVector<Scalar>*           targ_multi_vecs[]
 		,RTOpPack::ReductTarget         *reduct_obj
 		,const Index                    primary_first_ele
@@ -600,9 +600,9 @@ template<class Scalar>
 inline
 void applyOp(
 	const RTOpPack::RTOpT<Scalar>   &primary_op
-	,const size_t                   num_multi_vecs
+	,const int                   num_multi_vecs
 	,const MultiVector<Scalar>*     multi_vecs[]
-	,const size_t                   num_targ_multi_vecs
+	,const int                   num_targ_multi_vecs
 	,MultiVector<Scalar>*           targ_multi_vecs[]
 	,RTOpPack::ReductTarget*        reduct_objs[]
 	,const Index                    primary_first_ele      = 1
@@ -639,9 +639,9 @@ inline
 void applyOp(
 	const RTOpPack::RTOpT<Scalar>   &primary_op
 	,const RTOpPack::RTOpT<Scalar>  &secondary_op
-	,const size_t                   num_multi_vecs
+	,const int                   num_multi_vecs
 	,const MultiVector<Scalar>*     multi_vecs[]
-	,const size_t                   num_targ_multi_vecs
+	,const int                   num_targ_multi_vecs
 	,MultiVector<Scalar>*           targ_multi_vecs[]
 	,RTOpPack::ReductTarget         *reduct_obj
 	,const Index                    primary_first_ele      = 1
