+++ packages/TSFCore/src/interfaces/Core/TSFCoreVectorDecl.hpp
@@ -134,9 +134,9 @@ public:
 	 */
 	virtual void applyOp(
 		const RTOpPack::RTOpT<Scalar>    &op
-		,const size_t                    num_vecs
+		,const int                    num_vecs
 		,const Vector<Scalar>*           vecs[]
-		,const size_t                    num_targ_vecs
+		,const int                    num_targ_vecs
 		,Vector<Scalar>*                 targ_vecs[]
 		,RTOpPack::ReductTarget          *reduct_obj
 		,const Index                     first_ele
@@ -447,9 +447,9 @@ template<class Scalar>
 inline
 void applyOp(
 	const RTOpPack::RTOpT<Scalar>   &op
-	,const size_t                   num_vecs
+	,const int                   num_vecs
 	,const Vector<Scalar>*          vecs[]
-	,const size_t                   num_targ_vecs
+	,const int                   num_targ_vecs
 	,Vector<Scalar>*                targ_vecs[]
 	,RTOpPack::ReductTarget         *reduct_obj
 	,const Index                    first_ele     = 1
